- move logic from utils to ssz and wire modules. - ssz drops support for past forks
