+++ crates/pbs/src/mev_boost/status.rs
@@ -3,7 +3,7 @@ use std::time::{Duration, Instant};
 use axum::http::HeaderMap;
 use cb_common::{
     pbs::{RelayClient, error::PbsError},
-    utils::{get_user_agent_with_version, read_chunked_body_with_max},
+    wire::{get_user_agent_with_version, read_chunked_body_with_max},
 };
 use futures::future::select_ok;
 use reqwest::header::USER_AGENT;
