+++ tests/tests/pbs_mux.rs
@@ -12,7 +12,7 @@ use cb_common::{
     },
     signer::random_secret,
     types::Chain,
-    utils::{ResponseReadError, set_ignore_content_length},
+    wire::{ResponseReadError, set_ignore_content_length},
 };
 use cb_pbs::{DefaultBuilderApi, PbsService, PbsState};
 use cb_tests::{
@@ -358,7 +358,7 @@ async fn test_ssv_multi_with_node() -> Result<()> {
     info!("Sending get header");
     let res = mock_validator.do_get_header(Some(pubkey2.clone())).await?;
     assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(relay_state.received_get_header(), 1); // pubkey2 was loaded from the SSV node 
+    assert_eq!(relay_state.received_get_header(), 1); // pubkey2 was loaded from the SSV node
 
     // Shut down the server handles
     pbs_server.abort();
@@ -454,7 +454,7 @@ async fn test_ssv_multi_with_public() -> Result<()> {
     info!("Sending get header");
     let res = mock_validator.do_get_header(Some(pubkey2.clone())).await?;
     assert_eq!(res.status(), StatusCode::OK);
-    assert_eq!(relay_state.received_get_header(), 1); // pubkey2 was loaded from the SSV public API 
+    assert_eq!(relay_state.received_get_header(), 1); // pubkey2 was loaded from the SSV public API
 
     // Shut down the server handles
     pbs_server.abort();
